tests/wpt/tests/speculation-rules/prefetch/speculation-tags-no-tags.https.html

<!DOCTYPE html>
<title>Prefetch request's Sec-Speculation-Tags request headers</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.sub.js"></script>
<script src="resources/speculation-tags-utils.js"></script>
<script>
"use strict";

setup(() => assertSpeculationRulesIsSupported());

// https://wicg.github.io/nav-speculation/speculation-rules.html#collect-tags-for-matching-speculative-load-candidates
// https://wicg.github.io/nav-speculation/prefetch.html#sec-speculation-tags-header

// No tags case.
promise_test(async t => {
  const agent = await spawnWindow(t);
  const nextUrl = agent.getExecutorURL({ page: 2 });
  await agent.forceSpeculationRules({
      prefetch: [{source: "list", urls: [nextUrl]}]
  });
  await agent.navigate(nextUrl);

  const headers = await agent.getRequestHeaders();
  assert_prefetched(headers, "must be prefetched");
  assert_equals(headers.sec_speculation_tags, "null", "Sec-Speculation-Tags");
}, "Sec-Speculation-Tags: no tags");
</script>
